Teacher edit dashboard
