+++ apps/teacher/src/androidTest/java/com/instructure/teacher/ui/pages/CoursesListPage.kt
@@ -19,20 +19,12 @@ package com.instructure.teacher.ui.pages
 
 import android.view.View
 import androidx.recyclerview.widget.RecyclerView
-import androidx.test.espresso.Espresso.onView
 import androidx.test.espresso.action.ViewActions.click
 import androidx.test.espresso.contrib.RecyclerViewActions
-import androidx.test.espresso.matcher.ViewMatchers.withText
 import com.instructure.canvasapi2.models.Course
 import com.instructure.dataseeding.model.CourseApiModel
-import com.instructure.dataseeding.model.FavoriteApiModel
 import com.instructure.espresso.*
-import com.instructure.espresso.page.BasePage
-import com.instructure.espresso.page.callOnClick
-import com.instructure.espresso.page.onView
-import com.instructure.espresso.page.withAncestor
-import com.instructure.espresso.page.withId
-import com.instructure.espresso.page.withText
+import com.instructure.espresso.page.*
 import com.instructure.teacher.R
 import com.instructure.teacher.ui.utils.WaitForToolbarTitle
 import org.hamcrest.CoreMatchers
@@ -43,8 +35,6 @@ class CoursesListPage : BasePage() {
 
     private val toolbarTitle by WaitForToolbarTitle(R.string.courses)
 
-    private val menuEditFavoritesButton by WaitForViewWithId(R.id.menu_edit_favorite_courses)
-
     private val coursesTab by OnViewWithId(R.id.tab_courses)
 
     private val inboxTab by OnViewWithId(R.id.tab_inbox)
@@ -53,7 +43,7 @@ class CoursesListPage : BasePage() {
     private val coursesLabel by WaitForViewWithId(R.id.courseLabel)
 
     // Only displays if the user has favorite courses
-    private val seeAllCoursesLabel by WaitForViewWithId(R.id.seeAllTextView)
+    private val editDashboardLabel by WaitForViewWithId(R.id.editDashboardTextView)
 
     // Only displays if the user has no favorite courses
     private val emptyMessageLayout by WaitForViewWithId(R.id.emptyCoursesView)
@@ -66,7 +56,7 @@ class CoursesListPage : BasePage() {
 
     fun assertHasCourses(mCourses: List<Course>) {
         coursesLabel.assertDisplayed()
-        seeAllCoursesLabel.assertDisplayed()
+        editDashboardLabel.assertDisplayed()
 
         // Check that the recyclerview count matches course count
         coursesRecyclerView.check(RecyclerViewItemCountAssertion(mCourses.size))
@@ -87,8 +77,8 @@ class CoursesListPage : BasePage() {
         onView(matcher).assertDisplayed()
     }
 
-    fun openAllCoursesList() {
-        seeAllCoursesLabel.click()
+    fun openEditDashboard() {
+        editDashboardLabel.click()
     }
 
     fun openCourse(course: CourseApiModel) {
@@ -102,9 +92,5 @@ class CoursesListPage : BasePage() {
     fun openCourseAtPosition(position: Int) {
         // Add one to the position to account for the header in list
         coursesRecyclerView.perform(RecyclerViewActions.actionOnItemAtPosition<RecyclerView.ViewHolder>(position, click()))
-    }
-
-    fun openEditFavorites() {
-        menuEditFavoritesButton.click()
     }
 }
