+++ apps/teacher/src/androidTest/java/com/instructure/teacher/ui/pages/DashboardPage.kt
@@ -35,13 +35,12 @@ import org.hamcrest.Matcher
 class DashboardPage : BasePage() {
 
     private val toolbar by OnViewWithId(R.id.toolbar)
-    private val editFavoriteCourses by WaitForViewWithId(R.id.menu_edit_favorite_courses)
     private val coursesPageLabel by WaitForViewWithStringText("Courses")
     private val emptyView by OnViewWithId(R.id.emptyCoursesView, autoAssert = false)
     private val coursesView by OnViewWithId(R.id.swipeRefreshLayout, autoAssert = false)
     private val coursesHeaderWrapper by OnViewWithId(R.id.coursesHeaderWrapper, autoAssert = false)
     private val courseLabel by WaitForViewWithId(R.id.courseLabel)
-    private val seeAllCoursesButton by WaitForViewWithId(R.id.seeAllTextView)
+    private val editDashboardButton by WaitForViewWithId(R.id.editDashboardTextView)
     private val bottomBar by OnViewWithId(R.id.bottomBar)
     private val coursesTab by WaitForViewWithId(R.id.tab_courses)
     private val todoTab by WaitForViewWithId(R.id.tab_todo)
@@ -64,7 +63,7 @@ class DashboardPage : BasePage() {
         emptyView.assertNotDisplayed()
         onView(withParent(R.id.toolbar) + withText(R.string.courses)).assertDisplayed()
         coursesView.assertDisplayed()
-        seeAllCoursesButton.assertDisplayed()
+        editDashboardButton.assertDisplayed()
     }
 
     fun assertOpensCourse(course: CourseApiModel) {
@@ -73,12 +72,8 @@ class DashboardPage : BasePage() {
         onView(withId(R.id.courseBrowserTitle)).assertContainsText(course.name)
     }
 
-    fun clickSeeAll() {
-        onView(withId(R.id.seeAllTextView)).click()
-    }
-
-    fun openEditCoursesListPage() {
-        onView(withId(R.id.menu_edit_favorite_courses)).click()
+    fun clickEditDashboard() {
+        onView(withId(R.id.editDashboardTextView)).click()
     }
 
     fun openCourse(courseName: String) {
