+++ apps/teacher/src/androidTest/java/com/instructure/teacher/ui/utils/TeacherTest.kt
@@ -45,7 +45,6 @@ abstract class TeacherTest : CanvasTest() {
      * Required for auto complete of page objects within tests
      */
     val addMessagePage = AddMessagePage()
-    val allCoursesListPage = AllCoursesListPage()
     val announcementsListPage = AnnouncementsListPage()
     val assigneeListPage = AssigneeListPage()
     val assignmentDetailsPage = AssignmentDetailsPage()
@@ -68,7 +67,6 @@ abstract class TeacherTest : CanvasTest() {
     val discussionsListPage = DiscussionsListPage()
     val editAnnouncementPage = EditAnnouncementPage()
     val editAssignmentDetailsPage = EditAssignmentDetailsPage()
-    val editCoursesListPage = EditCoursesListPage()
     val editDiscussionsDetailsPage = EditDiscussionsDetailsPage()
     val editPageDetailsPage = EditPageDetailsPage()
     val editQuizDetailsPage = EditQuizDetailsPage()
